Code refactoring and code block support
