+++ brick-app/src/components/ChatInput.js
@@ -1,115 +1,11 @@
 import React, { useState } from 'react';
-import { makeStyles } from '@mui/styles';
 import TextField from '@mui/material/TextField';
 import Button from '@mui/material/Button';
 import SendIcon from '@mui/icons-material/Send';
 import InputAdornment from '@mui/material/InputAdornment';
 import IconButton from '@mui/material/IconButton';
 import { getConversations, getMessages, addMessageToConversation, addConversation, clearConversation } from '../ApiConnection';
-
-const useStyles = makeStyles((theme) => ({
-  container: {
-    display: 'flex',
-    flexDirection: 'column',
-    padding: theme.spacing(2),
-    height: 'calc(100% - 150px)',
-    marginTop: theme.spacing(2),
-    overflow: 'hidden',
-  },
-  input: {
-    flexGrow: 1,
-    marginBottom: theme.spacing(1),
-  },
-  responseBox: {
-    backgroundColor: theme.palette.background.paper,
-    margin: `${theme.spacing(1)}px 0`,
-    borderRadius: '10px',
-    overflowX: 'scroll',
-    fontFamily: 'Consolas, Monaco, "Andale Mono", "Ubuntu Mono", monospace',
-    fontSize: '0.9rem',
-    [theme.breakpoints.down('sm')]: {
-      maxWidth: '100%',
-    },
-  },
-  responseBoxCode: {
-    backgroundColor: theme.palette.background.paper,
-    padding: theme.spacing(2),
-    borderRadius: '10px',
-    marginTop: theme.spacing(1),
-    overflowX: 'scroll',
-    fontFamily: 'Consolas, Monaco, "Andale Mono", "Ubuntu Mono", monospace',
-    fontSize: '0.9rem',
-    [theme.breakpoints.down('sm')]: {
-      whiteSpace: 'pre-wrap',
-      wordBreak: 'break-all',
-    },
-    textAlign: 'left',
-  },
-  chatHistory: {
-    marginTop: theme.spacing(1),
-    border: `1px solid ${theme.palette.divider}`,
-    borderRadius: '10px',
-    padding: theme.spacing(2),
-    maxHeight: '300px',
-    overflowY: 'scroll',
-  },
-  chatHistoryEntry: {
-    marginBottom: theme.spacing(1),
-  },
-  errorMessage: {
-    color: 'red',
-  },
-  toggleHistoryButton: {
-    marginTop: theme.spacing(1),
-  },
-  clearConversationButton: {
-    marginTop: theme.spacing(1),
-  },
-  conversationButton: {
-    backgroundColor: 'white',
-    color: 'black',
-    border: '1px solid black',
-    padding: '8px',
-    borderRadius: '4px',
-    cursor: 'pointer',
-    margin: '4px',
-    '&.focused': {
-      backgroundColor: 'blue',
-      color: 'white',
-    },
-  },
-  code: {
-    backgroundColor: (theme) =>
-      theme.palette && theme.palette.mode === "dark"
-        ? "rgb(49, 49, 49)"
-        : "rgb(248, 248, 248)",
-    color: (theme) =>
-      theme.palette && theme.palette.mode === "dark"
-        ? "rgb(189, 147, 249)"
-        : "rgb(32, 32, 32)",
-    padding: "0.2rem 0.5rem",
-    borderRadius: "5px",
-    fontSize: "1rem",
-    whiteSpace: "pre-wrap",
-    overflowWrap: "break-word",
-    wordWrap: "break-word",
-    boxShadow: "1px 1px 5px rgba(0, 0, 0, 0.2)",
-    textAlign: 'left',
-  },
-  themedButton: {
-    backgroundColor: 'rgba(106, 27, 154, 0.12)',
-    color: theme.palette.mode === 'dark' ? 'white' : 'rgba(106, 27, 154, 0.87)',
-    border: `1px solid rgba(106, 27, 154, 0.42)`,
-    padding: '8px',
-    borderRadius: '4px',
-    cursor: 'pointer',
-    margin: '4px',
-    '&.focused': {
-      backgroundColor: 'rgba(106, 27, 154, 0.5)',
-      color: 'white',
-    },
-  },  
-}));
+import { useStyles } from '../css/components/ChatInput';
 
 function ChatInput({ onNewMessage, onTriggerImageGeneration }) {
   const [message, setMessage] = useState('');
@@ -121,12 +17,6 @@ function ChatInput({ onNewMessage, onTriggerImageGeneration }) {
   const [conversations, setConversations] = useState([]);
 
   const classes = useStyles();
-
-  const isCode = (text) => {
-    const codeIndicators = ['{', '}', ';', '(', ')', '=', '=>', 'function', 'const', 'let', 'var', 'import', 'export'];
-
-    return codeIndicators.some((indicator) => text.includes(indicator));
-  };
 
   const handleSubmit = async (event) => {
     event.preventDefault();
@@ -281,15 +171,7 @@ function ChatInput({ onNewMessage, onTriggerImageGeneration }) {
         />
       </form>
       {error && <p className={classes.errorMessage}>Error: {error}</p>}
-      {response && (
-        isCode(response) ? (
-          <pre className={classes.responseBoxCode}>
-            {response}
-          </pre>
-        ) : (
-          <p className={classes.responseBox}>{response}</p>
-        )
-      )}
+      {response && (<p className={classes.responseBox}>{response}</p>)}
       <div>
         <button
           className={classes.themedButton}
@@ -308,9 +190,8 @@ function ChatInput({ onNewMessage, onTriggerImageGeneration }) {
           <button
             key={index}
             onClick={() => handleGetConversationMessages(conversation, index)}
-            className={`${classes.themedButton} ${
-              currentConversation.index === index ? 'focused' : ''
-            }`}
+            className={`${classes.themedButton} ${currentConversation.index === index ? 'focused' : ''
+              }`}
           >
             {conversation.topic}
           </button>
@@ -322,11 +203,25 @@ function ChatInput({ onNewMessage, onTriggerImageGeneration }) {
       {showChatHistory && (
         <div className={classes.chatHistory}>
           <h1>Topic: {currentConversation.name}</h1>
-          {chatHistory.map((entry, index) => (
-            <div key={index} className={classes.chatHistoryEntry}>
-              <p>{entry.role.charAt(0).toUpperCase() + entry.role.slice(1)}: {entry.content}</p>
-            </div>
-          ))}
+          {chatHistory.map((entry, index) => {
+            const splitContent = entry.content.split('```');
+            return (
+              <div key={index} className={classes.chatHistoryEntry}>
+                {entry.role.charAt(0).toUpperCase() + entry.role.slice(1)}:{' '}
+                {splitContent.map((text, i) => {
+                  if (i % 2 === 1) {
+                    return (
+                      <p key={i} className={classes.codeBlock}>
+                        {text}
+                      </p>
+                    );
+                  } else {
+                    return text;
+                  }
+                })}
+              </div>
+            );
+          })}
         </div>
       )}
       {conversations.length > 0 && currentConversation.id !== '' && (
@@ -337,7 +232,7 @@ function ChatInput({ onNewMessage, onTriggerImageGeneration }) {
         </div>
       )}
     </div>
-  );  
+  );
 }
 
 function createNewMessage(message, isUser) {
